CBufferPayloadWriter moved to internal API (#1174)
* CBufferPayloadWriter moved to internal API (#1173)
